spt-2017/jobs-frontend-svc/requirements.txt
bottle==0.12.7
pymysql==0.7.2
pycrypto==2.6.1
python-jwt==2.0.1
tweepy==3.5.0
